requirements.txt
torch==1.12.1
tensorboard==2.12.0
scipy==1.10.1
tqdm==4.65.0
opencv-python==4.7.0.72
pillow==10.0.0